test_scenario_generator/requirements.txt
groq
python-docx
PyPDF2
argparse
gunicorn 
flask
pdfminer.six
requests
pdfplumber